Trigger event when refresh main window
